+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Version 0.7.0 (2016-10-07)
+
+  - feat: options.returns_count
+  - feat: support metadata & watch functionalities
+  - feat: npm packaging
+  - BREAKING: make Camomile a class
+
+## Version 0.6.3 (2015-06-28)
+
+  - feat: support for .../count routes
+
 ## Version 0.6.2 (2015-06-08)
 
   - feat: support for new non-destructive queue picking routes